changed floats to double to increase jade compability

+++ monitoring_bridge/src/monitoringtodiagnostics.cpp
@@ -80,7 +80,7 @@ int main(int argc, char **argv) {
 	ros::NodeHandle n("~");
 
 
-	float freq = 1;
+  double freq = 1;
 	if (!n.getParam("frequency", freq)) {
 		ROS_WARN("No frequency supplied. Working with %f Hz.", freq);
 	}
