+++ monitoring_fdir/src/recovery/recoveryExample.cpp
@@ -8,7 +8,7 @@ int main(int argc, char **argv)
   ros::init(argc, argv, "recoveryExample");
   ros::NodeHandle n("~");
 
-  float freq = 1;
+  double freq = 1;
   if (!n.getParam("frequency", freq))
   {
     freq = 1;
